add simplified video recorder, specify dtype in XmlBasedRobot to remove warning, clean up a bit

+++ Environments/pybullet_evo/robot_bases.py
@@ -21,9 +21,9 @@ class XmlBasedRobot:
     self.ordered_joints = None
     self.robot_body = None
 
-    high = np.ones([action_dim])
+    high = np.ones([action_dim], dtype=np.float32)
     self.action_space = gym.spaces.Box(-high, high)
-    high = np.inf * np.ones([obs_dim])
+    high = np.inf * np.ones([obs_dim], dtype=np.float32)
     self.observation_space = gym.spaces.Box(-high, high)
 
     #self.model_xml = model_xml
